ai-concierge-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
Faker==37.5.3
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
